+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### add
+
+- scheduled-reports
+
 ## [0.0.36] - 2026-05-26
 
 ### Fix